requirements.txt
numpy
pandas
matplotlib
scipy
opencv-python
ultralytics
supervision
inference
ttach
roboflow
tqdm
pyyaml